sq - move services to a services folder
