+++ src/main/java/hello/AuthControllerAdvice.java
@@ -10,6 +10,8 @@ import org.springframework.security.oauth2.client.authentication.OAuth2Authentic
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 
+import hello.services.*;
+
 @ControllerAdvice
 public class AuthControllerAdvice {
     @Autowired
