+++ src/main/java/hello/controllers/UsersController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 
 import hello.entities.AppUser;
 import hello.repositories.UserRepository;
-
+import hello.services.*;
 @Controller
 public class UsersController {
