+++ src/main/java/hello/services/EarthquakeQueryService.java
@@ -1,4 +1,4 @@
-package hello;
+package hello.services;
 
 
 import java.util.Arrays;
